@@ -4,7 +4,6 @@ import com.restaurant.kitchen.model.SimulationMode;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -16,7 +15,11 @@ public class Chef implements Runnable {
 
     private static final String[] CHEF_NAMES = {"Gordon", "Julia", "Marco", "Heston"};
 
-    private static final List<String> DISHES = Arrays.asList(
+    /**
+     * Direct Java 11 -> 21 migration:
+     * Arrays.asList() replaced with List.of() - returns a truly immutable list.
+     */
+    private static final List<String> DISHES = List.of(
             "Spaghetti Carbonara",
             "Mushroom Risotto",
             "Grilled Salmon",
@@ -69,6 +72,11 @@ public class Chef implements Runnable {
         return DISHES.get(random.nextInt(DISHES.size()));
     }
 
+    /**
+     * Direct Java 11 -> 21 migration:
+     * - if-else chain replaced with switch expression using arrow syntax (Java 14+)
+     * - Switch expressions are exhaustive: compiler ensures all enum values are handled
+     */
     private void processOrder() throws InterruptedException {
         String dish = pickDish();
         ReentrantLock stoveLock = stove.getLock();
@@ -76,15 +84,55 @@ public class Chef implements Runnable {
 
         logger.info("[Chef {}] New order received: {}. Heading to the kitchen...", name, dish);
 
-        if (mode == SimulationMode.DEADLOCK) {
-            if (id % 2 == 0) {
-                // Even chef: grabs Stove first, then Blender
-                logger.info("[Chef {}] Walking towards the STOVE to start cooking {}...", name, dish);
+        switch (mode) {
+            case DEADLOCK -> {
+                if (id % 2 == 0) {
+                    logger.info("[Chef {}] Walking towards the STOVE to start cooking {}...", name, dish);
+                    stoveLock.lockInterruptibly();
+                    logger.info("[Chef {}] Got the STOVE! Now I need the BLENDER for {}...", name, dish);
+                    try {
+                        blenderLock.lockInterruptibly();
+                        logger.info("[Chef {}] Got the BLENDER too! Making {} now!", name, dish);
+                        try {
+                            stove.use();
+                            blender.use();
+                            int total = ordersCompleted.incrementAndGet();
+                            logger.info("[Chef {}] Ding ding! {} is READY! (Orders served today: {})", name, dish, total);
+                        } finally {
+                            blenderLock.unlock();
+                        }
+                    } finally {
+                        stoveLock.unlock();
+                        logger.info("[Chef {}] Released STOVE and BLENDER. Ready for next order!", name);
+                    }
+                } else {
+                    logger.info("[Chef {}] Walking towards the BLENDER to prep {}...", name, dish);
+                    blenderLock.lockInterruptibly();
+                    logger.info("[Chef {}] Got the BLENDER! Now I need the STOVE for {}...", name, dish);
+                    try {
+                        stoveLock.lockInterruptibly();
+                        logger.info("[Chef {}] Got the STOVE too! Making {} now!", name, dish);
+                        try {
+                            blender.use();
+                            stove.use();
+                            int total = ordersCompleted.incrementAndGet();
+                            logger.info("[Chef {}] Ding ding! {} is READY! (Orders served today: {})", name, dish, total);
+                        } finally {
+                            stoveLock.unlock();
+                        }
+                    } finally {
+                        blenderLock.unlock();
+                        logger.info("[Chef {}] Released STOVE and BLENDER. Ready for next order!", name);
+                    }
+                }
+            }
+            case SAFE -> {
+                logger.info("[Chef {}] Walking towards the STOVE first (safe protocol)...", name);
                 stoveLock.lockInterruptibly();
-                logger.info("[Chef {}] Got the STOVE! Now I need the BLENDER for {}...", name, dish);
+                logger.info("[Chef {}] Got the STOVE! Now grabbing the BLENDER...", name);
                 try {
                     blenderLock.lockInterruptibly();
-                    logger.info("[Chef {}] Got the BLENDER too! Making {} now!", name, dish);
+                    logger.info("[Chef {}] Got the BLENDER! Cooking {} now!", name, dish);
                     try {
                         stove.use();
                         blender.use();
@@ -97,46 +145,6 @@ public class Chef implements Runnable {
                     stoveLock.unlock();
                     logger.info("[Chef {}] Released STOVE and BLENDER. Ready for next order!", name);
                 }
-            } else {
-                // Odd chef: grabs Blender first, then Stove (OPPOSITE ORDER = DEADLOCK!)
-                logger.info("[Chef {}] Walking towards the BLENDER to prep {}...", name, dish);
-                blenderLock.lockInterruptibly();
-                logger.info("[Chef {}] Got the BLENDER! Now I need the STOVE for {}...", name, dish);
-                try {
-                    stoveLock.lockInterruptibly();
-                    logger.info("[Chef {}] Got the STOVE too! Making {} now!", name, dish);
-                    try {
-                        blender.use();
-                        stove.use();
-                        int total = ordersCompleted.incrementAndGet();
-                        logger.info("[Chef {}] Ding ding! {} is READY! (Orders served today: {})", name, dish, total);
-                    } finally {
-                        stoveLock.unlock();
-                    }
-                } finally {
-                    blenderLock.unlock();
-                    logger.info("[Chef {}] Released STOVE and BLENDER. Ready for next order!", name);
-                }
-            }
-        } else {
-            // SAFE: All chefs take Stove -> Blender (consistent order, no deadlock)
-            logger.info("[Chef {}] Walking towards the STOVE first (safe protocol)...", name);
-            stoveLock.lockInterruptibly();
-            logger.info("[Chef {}] Got the STOVE! Now grabbing the BLENDER...", name);
-            try {
-                blenderLock.lockInterruptibly();
-                logger.info("[Chef {}] Got the BLENDER! Cooking {} now!", name, dish);
-                try {
-                    stove.use();
-                    blender.use();
-                    int total = ordersCompleted.incrementAndGet();
-                    logger.info("[Chef {}] Ding ding! {} is READY! (Orders served today: {})", name, dish, total);
-                } finally {
-                    blenderLock.unlock();
-                }
-            } finally {
-                stoveLock.unlock();
-                logger.info("[Chef {}] Released STOVE and BLENDER. Ready for next order!", name);
             }
         }
     }
