@@ -9,9 +9,12 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.PreDestroy;
-import java.lang.management.ManagementFactory;
-import java.lang.management.ThreadMXBean;
+/**
+ * Direct Java 11 -> 21 migration:
+ * javax.annotation.PreDestroy -> jakarta.annotation.PreDestroy
+ * Spring Boot 3.x migrated from Java EE (javax) to Jakarta EE (jakarta) namespace.
+ */
+import jakarta.annotation.PreDestroy;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.*;
@@ -33,6 +36,14 @@ public class KitchenSimulator {
     private SimulationMode mode = SimulationMode.DEADLOCK;
     private int chefCount = 4;
     private final List<String> chefNames = new ArrayList<>();
+    /**
+     * Direct Java 11 -> 21 migration:
+     * Progress-based deadlock detection fields. ThreadMXBean.findDeadlockedThreads()
+     * cannot detect deadlocks involving virtual threads (Java 21). Instead we monitor
+     * if orders stop progressing over a time window.
+     */
+    private int lastOrderCount = 0;
+    private long lastCheckTime = 0;
 
     public KitchenSimulator(Stove stove, Blender blender) {
         this.stove = stove;
@@ -48,31 +59,49 @@ public class KitchenSimulator {
         this.ordersCompleted.set(0);
         this.runningFlag = new AtomicBoolean(true);
         this.chefNames.clear();
-        this.executor = Executors.newFixedThreadPool(chefCount);
+        /**
+         * Direct Java 11 -> 21 migration:
+         * Executors.newFixedThreadPool(chefCount) -> Executors.newVirtualThreadPerTaskExecutor()
+         *
+         * Virtual Threads (JEP 444) are lightweight threads managed by the JVM:
+         * - Platform threads: ~1MB stack each, limited by OS (thousands max)
+         * - Virtual threads: ~few KB each, millions possible
+         * Each chef now runs on its own virtual thread.
+         */
+        this.executor = Executors.newVirtualThreadPerTaskExecutor();
 
-        logger.info("========================================");
-        logger.info("  RESTAURANT KITCHEN SIMULATOR");
-        logger.info("  Mode  : {}", mode);
-        logger.info("  Chefs : {}", chefCount);
-        logger.info("========================================");
+        /**
+         * Direct Java 11 -> 21 migration:
+         * Multiple logger.info() calls replaced with text blocks (Java 15+).
+         * Text blocks use triple-quote """ syntax for multi-line strings.
+         */
+        logger.info("""
+        ========================================
+          RESTAURANT KITCHEN SIMULATOR
+          Mode  : {}
+          Chefs : {}
+        ========================================""", mode.name(), chefCount);
 
         if (mode == SimulationMode.DEADLOCK) {
-            logger.info("WARNING: Deadlock mode enabled!");
-            logger.info("  -> Even chefs grab STOVE first, then BLENDER");
-            logger.info("  -> Odd chefs grab BLENDER first, then STOVE");
-            logger.info("  -> This WILL cause a deadlock!");
+            logger.info("""
+            WARNING: Deadlock mode enabled!
+              -> Even chefs grab STOVE first, then BLENDER
+              -> Odd chefs grab BLENDER first, then STOVE
+              -> This WILL cause a deadlock!""");
         } else {
-            logger.info("Safe mode enabled.");
-            logger.info("  -> All chefs follow the same lock order: STOVE -> BLENDER");
-            logger.info("  -> No deadlock possible.");
+            logger.info("""
+              Safe mode enabled.
+              -> All chefs follow the same lock order: STOVE -> BLENDER
+              -> No deadlock possible.""");
         }
 
-        logger.info("========================================");
-        logger.info("  Opening the kitchen doors...");
-        logger.info("========================================");
+        logger.info("""
+        ========================================
+          Opening the kitchen doors...
+        ========================================""");
 
         for (int i = 0; i < chefCount; i++) {
-            Chef chef = new Chef(i, stove, blender, mode, ordersCompleted, runningFlag);
+            var chef = new Chef(i, stove, blender, mode, ordersCompleted, runningFlag);
             chefNames.add(chef.getName());
             executor.submit(chef);
         }
@@ -97,10 +126,11 @@ public class KitchenSimulator {
                 Thread.currentThread().interrupt();
             }
         }
-        logger.info("========================================");
-        logger.info("  KITCHEN CLOSED");
-        logger.info("  Total orders served : {}", ordersCompleted.get());
-        logger.info("========================================");
+        logger.info("""
+        ========================================
+          KITCHEN CLOSED
+          Total orders served : {}
+        ========================================""", ordersCompleted.get());
     }
 
     public KitchenStatus getStatus() {
@@ -128,11 +158,30 @@ public class KitchenSimulator {
         );
     }
 
+    /**
+     * Direct Java 11 -> 21 migration:
+     * Replaced ThreadMXBean.findDeadlockedThreads() with progress-based detection.
+     * ThreadMXBean only detects platform thread deadlocks, not virtual thread deadlocks.
+     * This monitors order progress - if no orders complete within 2 seconds, infer deadlock.
+     */
     private boolean detectDeadlock() {
         if (!running) return false;
-        ThreadMXBean bean = ManagementFactory.getThreadMXBean();
-        long[] deadlockedThreads = bean.findDeadlockedThreads();
-        return deadlockedThreads != null && deadlockedThreads.length > 0;
+        long now = System.currentTimeMillis();
+        int currentOrders = ordersCompleted.get();
+
+        if (lastCheckTime == 0) {
+            lastCheckTime = now;
+            lastOrderCount = currentOrders;
+            return false;
+        }
+        if (now - lastCheckTime > 2000) {
+            if (currentOrders == lastOrderCount) {
+                return true;
+            }
+            lastOrderCount = currentOrders;
+            lastCheckTime = now;
+        }
+        return false;
     }
 
     @PreDestroy
